backend/requirements.txt
Flask
Flask-Cors
google-cloud-aiplatform
python-dotenv
google-generativeai
